# docs/swagger.yaml
definitions:
...
  gps.DistrictResponse:
    properties:
      district_id:
        example: 1
        type: integer
      district_name:
        example: เขต พระนคร
        type: string
    type: object
...
  gps.ProvinceResponse:
    properties:
      province_id:
        example: 1
        type: integer
      province_name:
        example: กรุงเทพมหานคร
        type: string
    type: object
  gps.SubDistrictResponse:
    properties:
      sub_district_id:
        example: 1
        type: integer
      sub_district_name:
        example: แขวง วัดอรุณ
        type: string
      zip_code:
        example: "10200"
        type: string
    type: object
...
  responses.BadRequestResponseSwagger:
    properties:
      message:
        description: ข้อความสถานะของ response
        example: Bad Request
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 400
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: false
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
...
  responses.InternalServerErrorResponseSwagger:
    properties:
      message:
        description: ข้อความสถานะของ response
        example: Internal Server Error
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 500
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: false
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
...
  responses.SuccessResponseSwagger-array_gps_DistrictResponse:
    properties:
      data:
        description: ข้อมูลที่ต้องการส่งกลับ
        items:
          $ref: '#/definitions/gps.DistrictResponse'
        type: array
        x-order: "4"
      message:
        description: ข้อความสถานะของ response
        example: OK
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 200
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: true
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
...
  responses.SuccessResponseSwagger-array_gps_ProvinceResponse:
    properties:
      data:
        description: ข้อมูลที่ต้องการส่งกลับ
        items:
          $ref: '#/definitions/gps.ProvinceResponse'
        type: array
        x-order: "4"
      message:
        description: ข้อความสถานะของ response
        example: OK
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 200
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: true
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
  responses.SuccessResponseSwagger-array_gps_SubDistrictResponse:
    properties:
      data:
        description: ข้อมูลที่ต้องการส่งกลับ
        items:
          $ref: '#/definitions/gps.SubDistrictResponse'
        type: array
        x-order: "4"
      message:
        description: ข้อความสถานะของ response
        example: OK
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 200
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: true
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
...
  responses.UnauthorizedResponseSwagger:
    properties:
      message:
        description: ข้อความสถานะของ response
        example: Unauthorized
        type: string
        x-order: "3"
      status:
        description: รหัสสถานะของ response
        example: 401
        type: integer
        x-order: "1"
      success:
        description: สถานะของ response
        example: false
        type: boolean
        x-order: "0"
      timestamp:
        description: เวลาที่ response ถูกสร้างขึ้น
        example: "2025-01-01T00:00:00.000000000Z"
        type: string
        x-order: "2"
    type: object
...
info:
  contact: {}
  description: For eZVeiw Lite Back Office Project.
  title: eZView Lite Back Office API
  version: "1.0"
paths:
...
  /v1/districts:
    get:
      consumes:
      - application/json
      description: |-
        API นี้ใช้สำหรับดึงรายชื่อเขตหรืออำเภอตามจังหวัดที่กำหนด

        - **provinceId**: รหัสจังหวัด
      parameters:
      - default: th
        description: เลือกภาษา (th หรือ en)
        enum:
        - th
        - en
        in: query
        name: lang
        type: string
      - description: รหัสจังหวัด
        in: query
        name: province_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: การร้องขอข้อมูลเขตหรืออำเภอสำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-array_gps_DistrictResponse'
        "400":
          description: คำขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: ดึงข้อมูลเขตหรืออำเภอตามจังหวัดที่เลือก
      tags:
      - Dropdowns
...
  /v1/provinces:
    get:
      consumes:
      - application/json
      description: API นี้ใช้สำหรับดึงรายชื่อจังหวัดตามภาษาที่กำหนด (th หรือ en)
      parameters:
      - default: th
        description: เลือกภาษา (th หรือ en)
        enum:
        - th
        - en
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: การร้องขอข้อมูลจังหวัดสำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-array_gps_ProvinceResponse'
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: ดึงข้อมูลจังหวัดตามภาษาที่เลือก
      tags:
      - Dropdowns
...
  /v1/subdistricts:
    get:
      consumes:
      - application/json
      description: |-
        API นี้ใช้สำหรับดึงรายชื่อตำบลหรือแขวงตามเขตหรืออำเภอที่กำหนด

        - **district_id**: รหัสเขตหรืออำเภอ
      parameters:
      - default: th
        description: เลือกภาษา (th หรือ en)
        enum:
        - th
        - en
        in: query
        name: lang
        type: string
      - description: รหัสเขตหรืออำเภอ
        in: query
        name: district_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: การร้องขอข้อมูลตำบลหรือแขวงสำเร็จ
          schema:
            $ref: '#/definitions/responses.SuccessResponseSwagger-array_gps_SubDistrictResponse'
        "400":
          description: คำขอไม่ถูกต้อง
          schema:
            $ref: '#/definitions/responses.BadRequestResponseSwagger'
        "401":
          description: Token ไม่ถูกต้องหรือหมดอายุ
          schema:
            $ref: '#/definitions/responses.UnauthorizedResponseSwagger'
        "500":
          description: ข้อผิดพลาดภายในเซิร์ฟเวอร์
          schema:
            $ref: '#/definitions/responses.InternalServerErrorResponseSwagger'
      security:
      - Bearer: []
      summary: ดึงข้อมูลตำบลหรือแขวงตามเขตหรืออำเภอที่เลือก
      tags:
      - Dropdowns
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"